modified

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -5,12 +5,13 @@
  */
 void print_rev(char *s)
 {
-	int len = 0, index;
+	long len = 0, index;
 
 	while (s[index++])
 		len++;
 
 	for (index = len - 1; index >= 0; index--)
 		_putchar(s[index]);
+
 	_putchar('\n');
 }
